move ES Mamager to utils
